# backend/bin/main/static/openapi/pesaguard-developer-v1.yaml
openapi: 3.1.0
info:
  title: PesaGuard Developer Platform API
  version: 0.1.0
  description: Implemented secure developer-platform foundation. OAuth, billing, production environments, webhooks, and financial execution are out of scope.
servers:
  - {url: 'http://localhost:8080', description: Local development}
security:
  - {bearerAuth: []}
paths:
...
  /api/v1/projects/{projectId}/environments:
    post: {parameters: [{$ref: '#/components/parameters/ProjectId'}], requestBody: {$ref: '#/components/requestBodies/CreateEnvironment'}, responses: {'200': {$ref: '#/components/responses/Data'}, '400': {$ref: '#/components/responses/Error'}, '403': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}}
    get: {parameters: [{$ref: '#/components/parameters/ProjectId'}], responses: {'200': {$ref: '#/components/responses/Data'}, '403': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}}}
  /api/v1/projects/{projectId}/environments/{environmentId}:
    get: {parameters: [{$ref: '#/components/parameters/ProjectId'}, {$ref: '#/components/parameters/EnvironmentId'}], responses: {'200': {$ref: '#/components/responses/Data'}, '403': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}}}
  /api/v1/projects/{projectId}/environments/{environmentId}/configuration:
    put: {parameters: [{$ref: '#/components/parameters/ProjectId'}, {$ref: '#/components/parameters/EnvironmentId'}], requestBody: {$ref: '#/components/requestBodies/UpdateEnvironmentConfiguration'}, responses: {'200': {$ref: '#/components/responses/Data'}, '400': {$ref: '#/components/responses/Error'}, '403': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}}
  /api/v1/projects/{projectId}/environments/{environmentId}/suspend:
    post: {parameters: [{$ref: '#/components/parameters/ProjectId'}, {$ref: '#/components/parameters/EnvironmentId'}], responses: {'200': {$ref: '#/components/responses/Data'}, '403': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}}
  /api/v1/projects/{projectId}/environments/{environmentId}/resume:
    post: {parameters: [{$ref: '#/components/parameters/ProjectId'}, {$ref: '#/components/parameters/EnvironmentId'}], responses: {'200': {$ref: '#/components/responses/Data'}, '403': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}}
  /api/v1/projects/{projectId}/environments/{environmentId}/deactivate:
    post: {parameters: [{$ref: '#/components/parameters/ProjectId'}, {$ref: '#/components/parameters/EnvironmentId'}], responses: {'200': {$ref: '#/components/responses/Data'}, '403': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}}}
  /api/v1/projects/{projectId}/environments/{environmentId}/promote:
    post: {parameters: [{$ref: '#/components/parameters/ProjectId'}, {$ref: '#/components/parameters/EnvironmentId'}], requestBody: {$ref: '#/components/requestBodies/PromoteEnvironment'}, responses: {'200': {$ref: '#/components/responses/Data'}, '400': {$ref: '#/components/responses/Error'}, '403': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}}
  /api/v1/projects/{projectId}/environments/{environmentId}/history:
    get: {parameters: [{$ref: '#/components/parameters/ProjectId'}, {$ref: '#/components/parameters/EnvironmentId'}], responses: {'200': {$ref: '#/components/responses/Data'}, '403': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}}}
  /api/v1/projects/{projectId}/environments/{environmentId}/api-keys:
    post: {parameters: [{$ref: '#/components/parameters/ProjectId'}, {$ref: '#/components/parameters/EnvironmentId'}], requestBody: {$ref: '#/components/requestBodies/CreateApiKey'}, responses: {'201': {$ref: '#/components/responses/KeyCreated'}, '400': {$ref: '#/components/responses/Error'}, '403': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}}}
    get: {parameters: [{$ref: '#/components/parameters/ProjectId'}, {$ref: '#/components/parameters/EnvironmentId'}], responses: {'200': {$ref: '#/components/responses/Data'}, '404': {$ref: '#/components/responses/Error'}}}
  /api/v1/projects/{projectId}/environments/{environmentId}/api-keys/{keyId}:
    delete: {parameters: [{$ref: '#/components/parameters/ProjectId'}, {$ref: '#/components/parameters/EnvironmentId'}, {$ref: '#/components/parameters/KeyId'}], responses: {'204': {description: Revoked}, '403': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}}
  /api/v1/projects/{projectId}/environments/{environmentId}/api-keys/{keyId}/suspend:
    post: {parameters: [{$ref: '#/components/parameters/ProjectId'}, {$ref: '#/components/parameters/EnvironmentId'}, {$ref: '#/components/parameters/KeyId'}], responses: {'200': {$ref: '#/components/responses/Data'}, '403': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}}
  /api/v1/projects/{projectId}/environments/{environmentId}/api-keys/{keyId}/resume:
    post: {parameters: [{$ref: '#/components/parameters/ProjectId'}, {$ref: '#/components/parameters/EnvironmentId'}, {$ref: '#/components/parameters/KeyId'}], responses: {'200': {$ref: '#/components/responses/Data'}, '403': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}}
  /api/v1/projects/{projectId}/environments/{environmentId}/api-keys/{keyId}/rotate:
    post: {parameters: [{$ref: '#/components/parameters/ProjectId'}, {$ref: '#/components/parameters/EnvironmentId'}, {$ref: '#/components/parameters/KeyId'}], responses: {'201': {$ref: '#/components/responses/KeyCreated'}, '403': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}}
  /api/v1/projects/{projectId}/environments/{environmentId}/api-keys/{keyId}/history:
    get: {parameters: [{$ref: '#/components/parameters/ProjectId'}, {$ref: '#/components/parameters/EnvironmentId'}, {$ref: '#/components/parameters/KeyId'}], responses: {'200': {$ref: '#/components/responses/Data'}, '403': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}}}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: opaque}
  parameters:
...
    ProjectId: {name: projectId, in: path, required: true, schema: {type: string, format: uuid}}
    EnvironmentId: {name: environmentId, in: path, required: true, schema: {type: string, format: uuid}}
    KeyId: {name: keyId, in: path, required: true, schema: {type: string, format: uuid}}
...
  requestBodies:
...
    CreateEnvironment: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CreateEnvironmentRequest'}}}}
...
    UpdateEnvironmentConfiguration: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/UpdateEnvironmentConfigurationRequest'}}}}
    PromoteEnvironment: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/PromoteEnvironmentRequest'}}}}
...
    CreateApiKey: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CreateApiKeyRequest'}}}}
...
  responses:
...
    KeyCreated: {description: One-time API key response; never cached, headers: {Cache-Control: {schema: {type: string, example: no-store}}}, content: {application/json: {schema: {$ref: '#/components/schemas/KeyEnvelope'}}}}
    Data: {description: Successful response envelope, content: {application/json: {schema: {$ref: '#/components/schemas/Envelope'}}}}
    Error: {description: Structured client error, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}}
...
  schemas:
...
    CreateEnvironmentRequest: {type: object, required: [name, type], properties: {name: {type: string, minLength: 2, maxLength: 80}, type: {type: string, enum: [DEVELOPMENT, SANDBOX, STAGING, PRODUCTION], description: A project may hold at most one environment per tier}}}
...
    UpdateEnvironmentConfigurationRequest: {type: object, required: [configuration], properties: {configuration: {type: object, description: Credential-like keys are rejected; secrets belong in environment credentials}}}
    PromoteEnvironmentRequest: {type: object, required: [targetType], properties: {targetType: {type: string, enum: [DEVELOPMENT, SANDBOX, STAGING, PRODUCTION], description: Must be exactly one tier above the current tier}, reason: {type: string, maxLength: 500}}}
...
    CreateApiKeyRequest: {type: object, required: [name, scopes, expiresIn], properties: {name: {type: string, minLength: 2, maxLength: 120}, scopes: {type: array, minItems: 1, maxItems: 20, uniqueItems: true, items: {type: string}}, expiresIn: {type: string, example: PT24H}}}
...
    Envelope: {type: object, required: [data, requestId], properties: {data: {type: object}, requestId: {type: string, format: uuid}}}
...
    KeyEnvelope: {type: object, required: [data, requestId], properties: {data: {type: object, required: [id, name, key, keyPrefix, scopes, expiresAt], properties: {id: {type: string, format: uuid}, name: {type: string}, key: {type: string, description: Returned only in this response}, keyPrefix: {type: string}, scopes: {type: array, items: {type: string}}, expiresAt: {type: string, format: date-time}}}, requestId: {type: string, format: uuid}}}
    ErrorEnvelope: {type: object, required: [error], properties: {error: {type: object, required: [code, message, requestId, timestamp, violations], properties: {code: {type: string}, message: {type: string}, requestId: {type: string, format: uuid}, timestamp: {type: string, format: date-time}, violations: {type: array, items: {type: object}}}}}}